requirements.txt
gspread>=6.2.0
google-auth>=2.40.0
